Remove triton, cleanup
