+++ src/butina.h
@@ -41,8 +41,6 @@ namespace nvMolKit {
  *                  centroid index for cluster i. Must be empty or size N.
  * @param reordering Whether to dynamically reorder candidates after each cluster assignment.
  * @param stream CUDA stream to execute operations on. Defaults to stream 0.
- * @param reordering Whether to dynamically reorder candidates after each cluster assignment.
- * @param stream CUDA stream to execute operations on. Defaults to stream 0.
  * @return Number of clusters assigned.
  */
 [[maybe_unused]] int butinaGpu(cuda::std::span<const double> distanceMatrix,
@@ -68,8 +66,6 @@ namespace nvMolKit {
  *                            processing of larger clusters but use more memory.
  * @param centroids Optional output array of size N. If provided, centroids[i] stores the
  *                  centroid index for cluster i. Must be empty or size N.
- * @param reordering Whether to dynamically reorder candidates after each cluster assignment.
- * @param stream CUDA stream to execute operations on. Defaults to stream 0.
  * @param reordering Whether to dynamically reorder candidates after each cluster assignment.
  * @param stream CUDA stream to execute operations on. Defaults to stream 0.
  * @return Number of clusters assigned.
